+++ src/state.rs
@@ -12,7 +12,7 @@ pub struct Config {
     pub fee_collector_address: CanonicalAddr,
     pub fee_collector_terrand: Decimal,
     pub terrand_address: CanonicalAddr,
-    pub live_round_max: u8,
+    pub live_round_max: u16,
 }
 pub const CONFIG: Item<Config> = Item::new("config");
